% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{call_models_parallel}
\alias{call_models_parallel}
\title{Call multiple models in parallel}
\usage{
call_models_parallel(models, prompt, temperature = NULL, timeout_seconds = 1000)
}
\arguments{
\item{models}{Vector of model names}

\item{prompt}{Text prompt to send to all models}

\item{temperature}{Sampling temperature (will be adjusted per model)}

\item{timeout_seconds}{Request timeout}
}
\value{
Named list of results (model_name -> response)
}
\description{
Call multiple models in parallel
}
