% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepcellseek_api_client.R
\name{get_deepcellseek_models}
\alias{get_deepcellseek_models}
\title{Get available models from DeepCellSeek API}
\usage{
get_deepcellseek_models(timeout_seconds = 30)
}
\arguments{
\item{timeout_seconds}{Request timeout}
}
\value{
List of available models by provider
}
\description{
Get available models from DeepCellSeek API
}
