% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_config.R
\name{get_model_config}
\alias{get_model_config}
\title{Get model configuration by model name}
\usage{
get_model_config(model)
}
\arguments{
\item{model}{Model name to lookup}
}
\value{
List containing model configuration or NULL if not found
}
\description{
Get model configuration by model name
}
