% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{get_model_provider}
\alias{get_model_provider}
\title{Get model provider from model name}
\usage{
get_model_provider(model)
}
\arguments{
\item{model}{Model name}
}
\value{
Provider name
}
\description{
Get model provider from model name
}
