% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_config.R
\name{get_supported_models}
\alias{get_supported_models}
\title{Get all supported models}
\usage{
get_supported_models()
}
\value{
Named list of all supported models grouped by provider
}
\description{
Get all supported models
}
