% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_config.R
\name{is_model_supported}
\alias{is_model_supported}
\title{Validate model exists}
\usage{
is_model_supported(model)
}
\arguments{
\item{model}{Model name to validate}
}
\value{
TRUE if model is supported, FALSE otherwise
}
\description{
Validate model exists
}
