% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepcellseek_api_client.R
\name{map_model_to_deepcellseek}
\alias{map_model_to_deepcellseek}
\title{Map model name to DeepCellSeek provider and model}
\usage{
map_model_to_deepcellseek(model)
}
\arguments{
\item{model}{Model name from our system}
}
\value{
List with provider and modelName
}
\description{
Map model name to DeepCellSeek provider and model
}
