% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_config.R
\docType{data}
\name{MODEL_CONFIGS}
\alias{MODEL_CONFIGS}
\title{Model Configuration Management}
\format{
An object of class \code{list} of length 7.
}
\usage{
MODEL_CONFIGS
}
\description{
This module manages the configuration of different LLM providers and models
to ensure easy maintenance and extensibility.
}
\keyword{datasets}
