% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepcellseek_api_client.R
\name{use_deepcellseek_api}
\alias{use_deepcellseek_api}
\title{Check if DeepCellSeek free API gateway should be used}
\usage{
use_deepcellseek_api()
}
\value{
Logical indicating if DeepCellSeek gateway should be used
}
\description{
Check if DeepCellSeek free API gateway should be used
}
